+++ packages/common/src/events/defaults.ts
@@ -49,6 +49,7 @@ export function buildDefaultEventRecord(): Partial<IEvent> {
     allowRegistration: true,
     attendanceType: [EventAttendanceType.IN_PERSON],
     categories: [],
+    inPersonCapacity: null,
     editGroups: [],
     endDateTime: endDateTime.toISOString(),
     notifyAttendees: true,