+++ packages/common/test/events/_internal/PropertyMapper.test.ts
@@ -31,23 +31,6 @@ describe("PropertyMapper", () => {
       end = new Date(start.valueOf() + 1000 * 60 * 60);
       eventRecord = {
         access: EventAccess.PRIVATE,
-        addresses: [
-          {
-            address: "1600 Pennsylvania Ave NW, Washington, DC 20500",
-            address2: "Suite 200",
-            capacity: 30,
-            createdAt: new Date().toISOString(),
-            description: "in-person description",
-            eventId: "31c",
-            extent: {},
-            geoAddress: "1600 Pennsylvania Ave NW, Washington, DC 20500",
-            geoAddrType: "something",
-            geoScore: 95,
-            location: {},
-            updatedAt: new Date().toISOString(),
-            venue: "The White House",
-          },
-        ],
         allDay: false,
         allowRegistration: false,
         attendanceType: [EventAttendanceType.IN_PERSON],
@@ -56,7 +39,15 @@ describe("PropertyMapper", () => {
         createdAt: now.toISOString(),
         createdById: "12345",
         creator: {
+          agoId: "abc",
           username: "jdoe",
+          email: "mockUser@gmail.com",
+          firstName: "mock",
+          lastName: "user",
+          deleted: false,
+          optedOut: false,
+          createdAt: now.toISOString(),
+          updatedAt: now.toISOString(),
         },
         description: "event description",
         editGroups: ["editGroup1"],
@@ -66,6 +57,7 @@ describe("PropertyMapper", () => {
         endTime: [end.getHours(), end.getMinutes(), end.getSeconds()].join(":"),
         endDateTime: end.toISOString(),
         geometry: null,
+        inPersonCapacity: 30,
         id: "31c",
         notifyAttendees: false,
         orgId: "42b",
@@ -169,7 +161,6 @@ describe("PropertyMapper", () => {
           url: "https://somewhere.com/",
         },
       ];
-      delete eventRecord.addresses;
       const res = propertyMapper.storeToEntity(eventRecord, {});
       expect(res.attendanceType).toEqual(HubEventAttendanceType.Online);
       expect(res.onlineCapacity).toEqual(20);
@@ -291,6 +282,7 @@ describe("PropertyMapper", () => {
         startTime: jasmine.any(String) as unknown as string,
         endDate: jasmine.any(String) as unknown as string,
         endTime: jasmine.any(String) as unknown as string,
+        inPersonCapacity: 30,
       } as IEvent);
     });
 
@@ -325,6 +317,7 @@ describe("PropertyMapper", () => {
         startTime: jasmine.any(String) as unknown as string,
         endDate: jasmine.any(String) as unknown as string,
         endTime: jasmine.any(String) as unknown as string,
+        inPersonCapacity: 30,
       } as IEvent);
     });
 
@@ -370,6 +363,7 @@ describe("PropertyMapper", () => {
         startTime: jasmine.any(String) as unknown as string,
         endDate: jasmine.any(String) as unknown as string,
         endTime: jasmine.any(String) as unknown as string,
+        inPersonCapacity: 30,
       } as IEvent);
     });
 
@@ -415,6 +409,7 @@ describe("PropertyMapper", () => {
         startTime: jasmine.any(String) as unknown as string,
         endDate: jasmine.any(String) as unknown as string,
         endTime: jasmine.any(String) as unknown as string,
+        inPersonCapacity: 30,
       } as IEvent);
     });
 
@@ -463,6 +458,7 @@ describe("PropertyMapper", () => {
         startTime: jasmine.any(String) as unknown as string,
         endDate: jasmine.any(String) as unknown as string,
         endTime: jasmine.any(String) as unknown as string,
+        inPersonCapacity: 30,
       } as IEvent);
     });
 
@@ -497,6 +493,7 @@ describe("PropertyMapper", () => {
         startTime: jasmine.any(String) as unknown as string,
         endDate: jasmine.any(String) as unknown as string,
         endTime: jasmine.any(String) as unknown as string,
+        inPersonCapacity: 30,
       } as IEvent);
     });
 
@@ -531,6 +528,7 @@ describe("PropertyMapper", () => {
         startTime: jasmine.any(String) as unknown as string,
         endDate: jasmine.any(String) as unknown as string,
         endTime: jasmine.any(String) as unknown as string,
+        inPersonCapacity: 30,
       } as IEvent);
     });
   });